fixed card's render fixed window's opening
