+++ src/components/Card.js
@@ -15,7 +15,7 @@ export class Card {
     this._selector = selector; //селектор контейнера куда вставлять карточки
     this._likes = data.likes; //данные с сервера о лайках на карточках
     this._id = data._id; //данные с сервера об id карточки
-    this._owner = data.owner._id; ////данные с сервера об id владельца карточки
+    this._ownerId = data.owner._id; ////данные с сервера об id владельца карточки
     this.userId = userId; // id пользователя
     this.apiDeleteLike = apiDeleteLike; // функция deleteLike с сервера,принимает 1 параметр, возвращает промис
     this.apiPutLike = apiPutLike; //функция putLike с сервера
@@ -58,15 +58,14 @@ export class Card {
     if (this._likes.some(obj => obj._id == this.userId)) {
       this._element.querySelector('.element__like-button').classList.add("element__like-button_active");
     }
-
     //если карточку создал не пользователь, на ней нет иконки удаления
-    if (this.owner._id !== this.userId) {
+    if (this._ownerId !== this.userId) {
       this._element.querySelector('.element__trash-button').remove();
     }
 
 
     //отрисовать количество лайков
-    this._element.querySelector('.element__like-counter').textContent = this.likes.length;
+    this._element.querySelector('.element__like-counter').textContent = this._likes._length;
 
     // Вернём элемент в качестве результата работы метода
     return this._element;
