--- src/components/utils.js
@@ -1,19 +1,0 @@
-export const cardsContainer = document.querySelector('.elements'); //контейнер для всех карточек
-
-export function hideInputError(formElement, inputElement, config) {
-  const errorElement = formElement.querySelector(`.${inputElement.id}-error`);
-  inputElement.classList.remove(config.inputErrorClass);
-  errorElement.classList.remove(config.errorClass);
-  errorElement.textContent = '';
-}
-
-export const validationConfig = {
-  formSelector: '.popup__input-container',
-  inputSelector: '.popup__input-item',
-  submitButtonSelector: '.popup__save-button',
-  inactiveButtonClass: 'popup__save-button_inactive',
-  inputErrorClass: 'popup__input-item_type_error',
-  errorClass: 'popup__input-item-error_active'
-};
-
-
